Wrapping script into app folder and extracting utils/mailing.
