+++ run.py
@@ -1,0 +1,5 @@
+from app.holyscrap import main
+
+
+if __name__ == '__main__':
+    main()
