+++ start.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-python holyscrap.py
+python run.py
